modified:   app/Http/Controllers/MapelController.php 	modified:   app/Http/Controllers/NilaiController.php 	modified:   resources/views/nilai/create.blade.php 	new file:   resources/views/nilai/edit.blade.php 	modified:   resources/views/nilai/index.blade.php 	modified:   routes/web.php

+++ resources/views/nilai/index.blade.php
@@ -3,7 +3,13 @@
     <center>
         <b>
             <h2>LIST NILAI</h2>
-            <a href="/nilai/create" class="button-primary">Tambah Data Nilai</a>
+            {{-- @if (session('user')->role == 'guru') --}}
+                <a href="/nilai/create" class="button-primary">Tambah Data Nilai</a>
+            {{-- @endif --}}
+
+            @if (session('success'))
+                <p class="text-success">{{ session('success') }}</p>
+            @endif
 
             <table>
                 <tr>
@@ -15,7 +21,9 @@
                     <th>UTS</th>
                     <th>UAS</th>
                     <th>NA</th>
-                    <th>ACTION</th>
+                    {{-- @if (session('user')->role == 'guru') --}}
+                        <th>ACTION</th>
+                    {{-- @endif --}}
                 </tr>
                 @foreach ($nilai as $n)
                     <tr>
@@ -27,10 +35,12 @@
                         <td>{{ $n->uts }}</td>
                         <td>{{ $n->uas }}</td>
                         <td>{{ $n->na }}</td>
-                        <td>
-                            <a href="/nilai/edit/{{ $n->id }}" class="button-warning">Edit</a>
-                            <a href="/nilai/destroy/{{ $n->id }}" class="button-danger">Hapus</a>
-                        </td>
+                        {{-- @if (session('user')->role == 'guru')     --}}
+                            <td>
+                                <a href="/nilai/edit/{{ $n->id }}" class="button-warning">Edit</a>
+                                <a href="/nilai/destroy/{{ $n->id }}" class="button-danger" onclick="return confirm('Yakin Hapus?')">Hapus</a>
+                            </td>
+                        {{-- @endif --}}
                     </tr>
                 @endforeach
             </table>
